+++ packages/ricos-plugins/web/src/plugin-text-button.ts
@@ -10,14 +10,68 @@ import type {
   ToolbarButtonProps,
   TranslationFunction,
   IContent,
+  ShortcutDataProvider,
+  Platform,
+  ModalService,
 } from 'ricos-types';
 import { resolversById } from 'wix-rich-content-toolbars-v3';
 
 export class PluginTextButton implements FormattingToolbarButton {
   private readonly button: FormattingToolbarButtonConfig;
 
-  constructor(button: FormattingToolbarButtonConfig) {
+  private readonly keyCombinationText: string;
+
+  private readonly t: TranslationFunction;
+
+  private readonly modalService: ModalService;
+
+  constructor(
+    button: FormattingToolbarButtonConfig,
+    modalService: ModalService,
+    shortcuts: ShortcutDataProvider,
+    t: TranslationFunction,
+    platform: Platform
+  ) {
     this.button = button;
+    this.t = t;
+    this.modalService = modalService;
+
+    this.keyCombinationText = shortcuts.getShortcutDisplayData(
+      button.id,
+      t,
+      platform
+    ).keyCombinationText;
+  }
+
+  private getTooltip() {
+    return this.button.presentation?.tooltip
+      ? `${this.t(this.button.presentation.tooltip)} ${
+          this.keyCombinationText ? `(${this.keyCombinationText})` : ''
+        }`
+      : '';
+  }
+
+  static of(
+    button: FormattingToolbarButtonConfig,
+    modalService: ModalService,
+    shortcuts: ShortcutDataProvider,
+    t: TranslationFunction,
+    platform: Platform
+  ) {
+    return new PluginTextButton(button, modalService, shortcuts, t, platform);
+  }
+
+  register() {
+    const modal = this.button.modal;
+    if (modal) {
+      this.modalService?.register(modal);
+    }
+  }
+
+  unregister() {
+    if (this.button.modal) {
+      this.modalService.unregister(this.button.modal?.id);
+    }
   }
 
   private toResolvedAttributes(): IToolbarItemConfigTiptap['attributes'] {
@@ -30,6 +84,10 @@ export class PluginTextButton implements FormattingToolbarButton {
   toToolbarItemConfig(editorCommands: EditorCommands): IToolbarItemConfigTiptap {
     return {
       ...this.button,
+      presentation: {
+        ...(this.button.presentation || {}),
+        tooltip: this.getTooltip(),
+      },
       attributes: this.toResolvedAttributes(),
       commands: {
         ...this.button.commands,
@@ -42,7 +100,7 @@ export class PluginTextButton implements FormattingToolbarButton {
 
   toExternalToolbarButtonConfig(
     editorCommands: EditorCommands,
-    t: TranslationFunction,
+    _t: TranslationFunction,
     _uploadService: IUploadService,
     _updateService: IUpdateService,
     content: IContent<unknown>
@@ -50,11 +108,18 @@ export class PluginTextButton implements FormattingToolbarButton {
     const attributes = this.toResolvedAttributes();
     return {
       type: 'button',
-      tooltip: t(this.button.presentation?.tooltip),
+      tooltip: this.getTooltip(),
       toolbars: [],
       getIcon: () => this.button.presentation?.icon,
       getLabel: () => this.button.id,
-      onClick: () => this.button.command?.(editorCommands),
+      onClick: () =>
+        this.button.modal
+          ? this.modalService?.isModalOpen(this.button.modal.id)
+            ? this.modalService?.closeModal(this.button.modal.id)
+            : this.modalService?.openModal(this.button.modal.id, {
+                layout: 'dialog',
+              })
+          : this.button.command?.(editorCommands),
       isActive: () => !!attributes.active && !!content.resolve(attributes.active),
       isDisabled: () => false,
       dataHook: this.button.presentation?.dataHook,
@@ -66,12 +131,12 @@ export class PluginTextButton implements FormattingToolbarButton {
 export class PluginTextButtons implements FormattingToolbarButtons {
   private readonly buttons: PluginTextButton[];
 
-  constructor(buttons: PluginTextButton[]) {
+  private constructor(buttons: PluginTextButton[]) {
     this.buttons = buttons;
   }
 
-  static of(buttons: FormattingToolbarButtonConfig[] = []): PluginTextButtons {
-    return new PluginTextButtons(buttons.map(button => new PluginTextButton(button)));
+  static of(buttons: PluginTextButton[]): PluginTextButtons {
+    return new PluginTextButtons(buttons);
   }
 
   asArray() {